moved class Util to api.general
